# package-search: improving performance on fetching facet display name (2nd try)