@davidread 2015-11-04
Closing in favour of slight rework with tests #2725 if that's ok?